+++ common/src/main/java/com/selfdot/cobblemontrainers/screen/TrainerPokemonScreen.java
@@ -4,12 +4,15 @@ import com.cobblemon.mod.common.CobblemonItems;
 import com.selfdot.cobblemontrainers.trainer.Trainer;
 import com.selfdot.cobblemontrainers.trainer.TrainerPokemon;
 import com.selfdot.cobblemontrainers.util.PokemonUtility;
+import com.selfdot.cobblemontrainers.util.ScreenUtils;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.inventory.Inventory;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 import net.minecraft.screen.slot.SlotActionType;
 import net.minecraft.text.Text;
+import net.minecraft.util.Formatting;
 
 public class TrainerPokemonScreen extends Screen {
 
@@ -23,6 +26,7 @@ public class TrainerPokemonScreen extends Screen {
     private int levelSlot;
     private int natureSlot;
     private int shinySlot;
+    private int heldItemSlot;
 
     public TrainerPokemonScreen(Trainer trainer, TrainerPokemon trainerPokemon) {
         super(new TrainerTeamScreen(trainer));
@@ -43,6 +47,7 @@ public class TrainerPokemonScreen extends Screen {
         levelSlot = evsSlot + columns;
         natureSlot = levelSlot + 1;
         shinySlot = levelSlot - 1;
+        heldItemSlot = levelSlot - 2;
 
         setSlot(inventory, movesSlot, Items.MUSIC_DISC_5, "Moves");
         setSlot(inventory, abilitiesSlot, CobblemonItems.CLOVER_SWEET, "Abilities");
@@ -52,6 +57,17 @@ public class TrainerPokemonScreen extends Screen {
         setSlot(inventory, shinySlot, Items.NETHER_STAR, "Toggle Shiny");
         setSlot(inventory, levelSlot, CobblemonItems.WISE_GLASSES, "Level");
         setSlot(inventory, natureSlot, Items.CYAN_DYE, "Nature");
+
+        Item heldItem = trainerPokemon.getHeldItem();
+        ItemStack heldItemItemStack = new ItemStack(heldItem.equals(Items.AIR) ? Items.STICK : heldItem);
+        heldItemItemStack.setCustomName(Text.literal("Held Item"));
+        ScreenUtils.addLore(heldItemItemStack, new Text[]{
+            Text.literal(heldItem.equals(Items.AIR) ?
+                Formatting.RED + "None" :
+                Formatting.GREEN + heldItem.getDefaultStack().getName().getString()
+            )
+        });
+        inventory.setStack(heldItemSlot, heldItemItemStack);
     }
 
     @Override
@@ -76,6 +92,8 @@ public class TrainerPokemonScreen extends Screen {
             trainerPokemon.toggleShiny();
             trainer.save();
             player.openHandledScreen(new TrainerSetupHandlerFactory(this));
+        } else if (slotIndex == heldItemSlot) {
+            player.openHandledScreen(new TrainerSetupHandlerFactory(new HeldItemSelectScreen(trainer, trainerPokemon)));
         }
     }