+++ common/src/main/java/com/selfdot/cobblemontrainers/trainer/TrainerPokemon.java
@@ -1,19 +1,36 @@
 package com.selfdot.cobblemontrainers.trainer;
 
+import com.cobblemon.mod.common.CobblemonItems;
+import com.cobblemon.mod.common.api.Priority;
 import com.cobblemon.mod.common.api.abilities.Abilities;
 import com.cobblemon.mod.common.api.abilities.Ability;
+import com.cobblemon.mod.common.api.events.CobblemonEvents;
 import com.cobblemon.mod.common.api.moves.MoveSet;
 import com.cobblemon.mod.common.api.moves.Moves;
 import com.cobblemon.mod.common.api.pokemon.Natures;
 import com.cobblemon.mod.common.api.pokemon.PokemonSpecies;
 import com.cobblemon.mod.common.pokemon.*;
+import com.cobblemon.mod.common.pokemon.properties.UncatchableProperty;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.selfdot.cobblemontrainers.util.DataKeys;
+import kotlin.Unit;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.item.Items;
+import net.minecraft.registry.Registries;
+import net.minecraft.registry.RegistryKeys;
 import net.minecraft.util.Identifier;
 
+import java.util.HashSet;
+import java.util.Set;
+import java.util.UUID;
+
 public class TrainerPokemon {
+
+    public static final Set<UUID> IS_TRAINER_OWNED = new HashSet<>();
+    public static final Set<UUID> MUST_REENABLE_LOOT_GAMERULE = new HashSet<>();
 
     private Species species;
     private Gender gender;
@@ -24,6 +41,9 @@ public class TrainerPokemon {
     private IVs ivs;
     private EVs evs;
     private boolean isShiny = false;
+    private Item heldItem = null;
+
+    private final UUID uuid = UUID.randomUUID();
 
     public TrainerPokemon() { }
 
@@ -48,6 +68,11 @@ public class TrainerPokemon {
             moveset.setMove(i, Moves.INSTANCE.getByName(movesetJson.get(i).getAsString()).create());
         }
         if (jsonObject.has(DataKeys.POKEMON_SHINY)) isShiny = jsonObject.get(DataKeys.POKEMON_SHINY).getAsBoolean();
+        if (jsonObject.has(DataKeys.POKEMON_HELD_ITEM)) {
+            heldItem = Registries.ITEM.get(
+                Identifier.tryParse(jsonObject.get(DataKeys.POKEMON_HELD_ITEM).getAsString())
+            );
+        }
     }
 
     public JsonObject toJson() {
@@ -63,6 +88,7 @@ public class TrainerPokemon {
         jsonObject.add(DataKeys.POKEMON_IVS, ivs.saveToJSON(new JsonObject()));
         jsonObject.add(DataKeys.POKEMON_EVS, evs.saveToJSON(new JsonObject()));
         jsonObject.addProperty(DataKeys.POKEMON_SHINY, isShiny);
+        jsonObject.addProperty(DataKeys.POKEMON_HELD_ITEM, Registries.ITEM.getId(heldItem).toString());
         return jsonObject;
     }
 
@@ -78,6 +104,11 @@ public class TrainerPokemon {
         ivs.spliterator().forEachRemaining(entry -> pokemon.setIV(entry.getKey(), entry.getValue()));
         evs.spliterator().forEachRemaining(entry -> pokemon.setEV(entry.getKey(), entry.getValue()));
         pokemon.setShiny(isShiny);
+        if (heldItem.equals(Items.AIR)) pokemon.removeHeldItem();
+        else pokemon.swapHeldItem(new ItemStack(heldItem), false);
+        pokemon.setUuid(uuid);
+        pokemon.getCustomProperties().add(UncatchableProperty.INSTANCE.uncatchable());
+        IS_TRAINER_OWNED.add(uuid);
         return pokemon;
     }
 
@@ -92,6 +123,7 @@ public class TrainerPokemon {
         trainerPokemon.ivs = pokemon.getIvs();
         trainerPokemon.evs = pokemon.getEvs();
         trainerPokemon.isShiny = pokemon.getShiny();
+        trainerPokemon.heldItem = pokemon.heldItem().getItem();
         return trainerPokemon;
     }
 
@@ -129,6 +161,23 @@ public class TrainerPokemon {
 
     public void toggleShiny() {
         isShiny = !isShiny;
+    }
+
+    public Item getHeldItem() {
+        return heldItem;
+    }
+
+    public void setHeldItem(Item heldItem) {
+        this.heldItem = heldItem;
+    }
+
+    public static void registerPokemonSendOutListener() {
+        CobblemonEvents.POKEMON_SENT_POST.subscribe(Priority.NORMAL, event -> {
+            if (IS_TRAINER_OWNED.contains(event.getPokemon().getUuid())) {
+                event.getPokemonEntity().getUnbattleable().set(true);
+            }
+            return Unit.INSTANCE;
+        });
     }
 
 }