+++ common/src/main/java/com/selfdot/cobblemontrainers/util/DataKeys.java
@@ -11,6 +11,7 @@ public class DataKeys {
     public static final String POKEMON_ABILITY = "ability";
     public static final String POKEMON_MOVESET = "moveset";
     public static final String POKEMON_SHINY = "shiny";
+    public static final String POKEMON_HELD_ITEM = "heldItem";
     public static final String TRAINER_NAME = "name";
     public static final String TRAINER_TEAM = "team";
     public static final String TRAINER_MONEY_REWARD = "moneyReward";
@@ -18,6 +19,7 @@ public class DataKeys {
     public static final String TRAINER_LOSS_COMMAND = "lossCommand";
     public static final String TRAINER_CAN_ONLY_BEAT_ONCE = "canOnlyBeatOnce";
     public static final String TRAINER_GROUP = "group";
+    public static final String TRAINER_COOLDOWN_SECONDS = "cooldownSeconds";
     public static final String UNGROUPED = "Ungrouped";
     public static final String PLAYER_TOKEN = "%player%";