+++ common/src/main/kotlin/com/selfdot/cobblemontrainers/CobblemonTrainers.kt
@@ -3,11 +3,8 @@ package com.selfdot.cobblemontrainers
 import com.mojang.brigadier.CommandDispatcher
 import com.mojang.logging.LogUtils
 import com.selfdot.cobblemontrainers.command.TrainerCommandTree
-import com.selfdot.cobblemontrainers.config.CobblemonConfig
 import com.selfdot.cobblemontrainers.screen.SpeciesSelectScreen
-import com.selfdot.cobblemontrainers.trainer.TrainerBattleListener
-import com.selfdot.cobblemontrainers.trainer.TrainerRegistry
-import com.selfdot.cobblemontrainers.trainer.TrainerWinTracker
+import com.selfdot.cobblemontrainers.trainer.*
 import com.selfdot.cobblemontrainers.util.CobblemonTrainersLog
 import dev.architectury.event.events.common.CommandRegistrationEvent
 import dev.architectury.event.events.common.LifecycleEvent
@@ -21,12 +18,10 @@ object CobblemonTrainers {
     const val MODID = "cobblemontrainers"
     val TRAINER_REGISTRY = TrainerRegistry(this)
     val TRAINER_WIN_TRACKER = TrainerWinTracker(this)
+    val TRAINER_COOLDOWN_TRACKER = TrainerCooldownTracker(this)
     private var disabled = false
     private val LOGGER = LogUtils.getLogger()
     fun initialize() {
-        // Load official Cobblemon's config.
-        CobblemonConfig()
-
         LifecycleEvent.SERVER_STARTING.register(CobblemonTrainers::onServerStart)
         LifecycleEvent.SERVER_STOPPING.register(CobblemonTrainers::onServerStop)
 
@@ -58,7 +53,10 @@ object CobblemonTrainers {
         CobblemonTrainersLog.LOGGER.info("Loading trainer data")
         TRAINER_REGISTRY.load()
         TRAINER_WIN_TRACKER.load()
-        TrainerBattleListener.getInstance().setServer(server);
+        TRAINER_COOLDOWN_TRACKER.load()
+        TrainerBattleListener.getInstance().setServer(server)
+        Generation5AI.initialiseTypeChart()
+        TrainerPokemon.registerPokemonSendOutListener()
     }
 
     private fun onServerStop(server: MinecraftServer) {
@@ -66,6 +64,7 @@ object CobblemonTrainers {
             CobblemonTrainersLog.LOGGER.info("Storing trainer data")
             TRAINER_REGISTRY.save()
             TRAINER_WIN_TRACKER.save()
+            TRAINER_COOLDOWN_TRACKER.save()
         }
     }